implemented payment getway stripe

+++ src/lib/cart/deleteItem.js
@@ -16,4 +16,19 @@ const deleteItem = async (id) => {
   }
 };
 
-module.exports = deleteItem;
+const deleteItemByUserId = async (id) => {
+  try {
+    const result = await Cart.deleteItemByUserId(id);
+
+    if (result.rowCount === 0) {
+      throw notFoundError("Cart not found!");
+    }
+
+    // Return the deleted user object
+    return result.rows[0];
+  } catch (error) {
+    throw error;
+  }
+};
+
+module.exports = {deleteItem, deleteItemByUserId};
