+++ src/lib/cart/index.js
@@ -2,7 +2,7 @@ const create = require("./create");
 const {findSingleItem, existCart, isExistUserCart}= require("./findSingleItem");
 const findAllItems = require("./findAllItems");
 const {updateItem, updateQuantity} = require("./updateItem");
-const deleteItem = require("./deleteItem");
+const {deleteItem, deleteItemByUserId} = require("./deleteItem");
 const checkOwnership = require("./ownership");
 const findAllItemsByUserId = require("./findAllItemsByUserId");
 
@@ -16,5 +16,6 @@ module.exports = {
   updateQuantity,
   deleteItem,
   checkOwnership,
-  findAllItemsByUserId
+  findAllItemsByUserId,
+  deleteItemByUserId
 };
